fix tour add code dashboard

+++ src/main/java/com/example/demo/repository/KoiFishOrderRepository.java
@@ -17,7 +17,7 @@ public interface KoiFishOrderRepository extends JpaRepository<KoiFishOrder,Long>
     Double findTotalOrderAmountByMonthAndYear(@Param("month") int month, @Param("year") int year);
 
     @Query("SELECT COUNT(k) FROM KoiFishOrder k WHERE k.isDeleted = false AND MONTH(k.createAt) = :month AND YEAR(k.createAt) = :year")
-    Long countOrdersByMonthAndYear(@Param("month") int month, @Param("year") int year);
+    Double countOrdersByMonthAndYear(@Param("month") int month, @Param("year") int year);
 
     @Query("SELECT COUNT(k) FROM KoiFishOrder k WHERE k.isDeleted = true AND MONTH(k.createAt) = :month AND YEAR(k.createAt) = :year")
     Long countDeletedOrdersByMonthAndYear(@Param("month") int month, @Param("year") int year);
